feat(oncall): add envoy specific healthcheck endpoint (#3695)
We are currently conflating two different healthchecks as one. There's the healthcheck that envoy uses to not send traffic to the pod while it's draining, and the healthcheck that k8s uses to know whether to restart the pod. Currently those are the same thing.

The hypothesis is that if we have them as two separate endpoints, we will stop getting erroneous restarts by k8s of the snuba API