fix: Reject improperly typed tag conditions as invalid queries
Validate that simple tag conditions (conditions with a tags column on the left
side and a literal or array of literals on the right side) are comparing against
strings and nothing else.

This is a common error that comes from Sentry, and other current checks don't
catch it so it gets sent to Clickhouse and reported to Sentry as an error.