feat(admin): allow querylog to be queried by more threads (#3718)
There are many useful querylog queries which we cannot do because we only have one thread to query the querylog with. Up the max_threads to 4 but make it runtime configurable.

NOTE:

I'm not sure how to test this to know that it works as expected (except for deploying it). It seems like you can pass anything you want into the `settings` dict, and have max_threads be any value and the query will still succeed. Would love some guidance on this if anyone has ideas.