fix: Check simple conditions on datetime columns
A common error that happens when querying Snuba is to compare a datetime column
to a literal value that isn't a datetime. Previously a query with this error
would run on Sentry and end up logging an exception. While this doesn't handle
all the edge cases, it does catch the most commonly seen use case.