feat(admin) Add role authentication to individual tools
In this change, a tool is currently linked to a tab in the frontend UI. Every
endpoint in the admin tool (minus the health check and a few others) has been
annotated with a tool name.

These tool names are then checked as resources in the given AdminUser, and will
return a 403 if a user doesn't have access to the particular tool or the
catchall resource that gives blanket access.

This policy means by default a user will see nothing, and the new resouces will
need to be added to everyone who currently has access to the admin tool.

This will allow whitelisting specific tools for users, so access to tools can
start being given to product teams.

All the members of SnS have been added to the iam_policy file with permissions
to all the tools. Once this is in place, we can start giving specific
permissions to users outside of the team as well.

This change also limits what is seen in the admin tool to the tools allowed.
This isn't done in a secure way, but since the endpoints are secure this is more
of a user experience change.