fix(oncall): error on non-existent columns (#3765)
Queries get sent to us with non-existent columns in them. No need to let those queries get all the way to clickhouse. Turn on the EntityContainsColumnsValidator to ERROR for all entities by default.

As part of this I moved the test for this validator to its own file and decoupled it from the outcomes entity (it now uses its a test configuration entity)