ref(ratelimit): Create separate namespace if config redis for rate limit keys
Motivation:

Currently all of the configuration for rate limiting is done alongside other
runtime configuration. This was fine when not much rate limiting was done, but
now rate limit keys are the dominant keys in the configuration tool.

Ultimately there should be a separate tool for configuring rate limiting. That
simplifies the user interface and allows product owners access to rate limits
without having to grant access to configuration in general.

Blast Radius:

This change adds a new config key hash to redis to store the rate limiting keys.
This still uses the configuration mechanism under the hood, largely to leverage
the audit log that state changes allow.

This is step one of the migration process, doing dual read/writes on the keys
when they are accessed. This will be followed up with a migration to move all
the old keys over automatically.

Notes:

The new rate limit keys will not show up in the config tool, so follow up
changes will be added to expose them.