feat(rust-consumer): Add yaml settings file
Adds a yaml file with the minimal settings needed to run the Rust consumer
(the Kafka broker config and the ClickHouse cluster mapping).

These values are essentially duplicated from the Python settings file. If Rust and
Python code were being run side by side, these would need to be kept in sync.
Though since only consumers would ever be in Rust, not every setting will be
duplicated here, just a few.

An alternative approach would be to attempt to load the Python file in Rust so that
the values would be defined only once. This option was not pursued as it is messier
and likely not the right long term direction for settings anyway. It has been a long
time goal in Snuba to move away from the Python specific settings to something language
agnostic that is easier for end users (including open source users) to configure. For this
reason, I wanted to propose starting fresh with yaml settings rather than attempting
to share the Python ones.