ref(CapMan): remove execute_query_strategy variable (#3795)
in order to make room for capacity management, we have to understand how db_query.py works. As part of that process, I realized that we had dead code in this file `execute_query_with_caching` has not been used since 2020 and we're not going to use it again.

@untitaker actually tried to remove this in #3372 but tests were failing. 

### Significant changes

added a short circuit option to the readthrough cache. This was done for the following reasons:

1. The test that was failing was failing because it was hitting the readthrough cache, which never hit the `execute_query` function, thus never setting the `consistent` flag in the results. 
2. We had a single tenant incident where we wanted to disable the reathrough cache and were not able to so I figured it was fine to add it. This way if we lose redis for a consistent amount of time we can just pass the queries right through.