ref: Simplify event schema, add kafka payload as sentry attachment
We are running into event payload size limitations when trying to
attach the "invalid" payload. Instead, try sentry attachments for this
purpose. Since this might be expensive (though it shouldn't be), it is
also now only set up if the sampling decision is true.

Also remove some constructs from event schema that Snuba does not use.