feat: Add new statuses for SLO (#3747)
This adds a new type of status that is more granular than the existing status.
The new status is added in parallel to the old one, so that the migration can
happen seemlessly without losing data. Once the old status isn't being used
anymore it can be removed and this code simplified.

The logic for determing the status has been consolidated to where the statuses
are defined to make it easier to keep track of the logic behind the different
values.

Once this is merged the corresponding SLOs can be set up in Datadog and checked.