feat: Share Python settings with Rust and add Rust consumer entrypoint (#3856)
Alt to https://github.com/getsentry/snuba/pull/3802

I prefer this approach at least in the short term as it keeps the settings in Rust aligned with the Python version. It requires more minimal changes to the Python settings than the other option: only that the timedelta types need to change as these are not easily json serializable. Whether Python settings should ultimately move to yaml is a larger, more consequential topic - let's not try to solve it within the Rust consumer prototype.

The Rust consumer entrypoint loads the Python settings, writes the json file and replaces the current process with the Rust program passing the path to the JSON file through to the Rust consumer.