feat: Start build the config for Rust consumer
We don't want to reimplement the slice/cluster resolution in Rust,
call it in Python when the Rust consumer starts and write it in
a file for Rust.

This is currently incomplete, there are some additional fields we need to add here
such as the reference to the message processor.