feat(rust-arroyo): Split the message interface to support batching (#3885)
Same as we did in Python Arroyo https://github.com/getsentry/arroyo/pull/134.

The processing strategy interface needs to support messages that represent a batch or a filter message - i.e. not necessarily a 1:1 relationship between messages from the Kafka broker and a message passed to the processing strategy. This means a message can be associated with a number of partitions and offsets, not just a single one.