feat: DLQ performance test
We want to check if the DLQ proposal in Arroyo is viable by testing
the performance impact on Snuba in production.

There are no functional changes here, however if the runtime config
`run_dlq_test_{storage_name}` is set, the consumer for that storage
will start buffering messages as would be needed for the DLQ.

Note that this config is only checked once upon start up of the consumer
so a deploy is needed for the runtime config value to take effect.