ref: Dispatch replacements based on get_replacement_type
Currently we first dispatch to replacement based on type based on a big
if-else statement, then ask the replacement what type it is. Either the
replacement has a say about the ReplacementType or it doesn't, but
we shouldn't use it as a databag for random things we want to
forward to another function.