ref: Start using FailoverRedis in Snuba
There is a class called FailoverRedis in Sentry's sentry.utils that is
supposed to deal with networking errors when GCP memorystore falls over.
Snuba doesn't have an equivalent class, and as a result would be
susceptible to the same failures.

sentry-redis-tools is a new package that contains this code + maybe more
redis-related code in the future. For now, it only contains a copy of
Sentry's FailoverRedis.

Add FailoverRedis and make it the default redis client for everything.

Sentry does NOT use FailoverRedis as the default redis client, but I
believe we should change that as well to prevent incidents like INC-311.