fix: Fix broken rust_snuba import
`import rust_snuba` was accidentally importing the folder in the
repository instead of the installed wheel. On top of that, installing
the wheel never happened because the arguments for `find` were in the
wrong order.

Now that that is fixed, we can no longer assume that rust_snuba is
available in all builds, therefore we remove the `import rust_snuba`
sanity check -- not like it helped us catch anything anyway.