feat: Kafka topic configuration (#4082)
We need to start aligning topic configuration and encoding the correct values
so that they match in all environments. 50MB is way too large and does not
reflect any actual values set in prod. This is problematic because we don't
see these issues in dev that actually exist in prod. Reduce the default to 10MB,
with the exception of Replays which is 15MB.