fix(migrations): Allow setting file changes to happen in migrations
When a new migration group is added, a key element is registering the storage set key.
Storage set keys can be hardcoded or be dynamically configured using the settings file.
In order to use the dynamically configured settings file, we should allow modifications
to settings file to happen in the same PR as migrations.