fix(consumers): Add timeout for http batch writes
When a join is being called on HTTPWriteBatch we should limit the time spent
waiting for the operation to be successful. If we do not limit the time spent
the consumer can stay hung for any amount of time. It would stay hung till
the kafka broker realizes that the consumer is dead (which today is 5 minutes).
Instead of waiting so long, we should fail fast and abort earlier.