ref(tech-debt) Column validator also checks mapped columns (#4116)
The EntityContainsColumnValidator wasn't checking the mapped columns when
validating if a column was allowed on the entity. Change this so that any mapped
columns are also marked as valid.

Also enable this on one dataset (search_issues) as a test. If everything works
correctly, the rest of the dataset will also get this validator.