remove extended allocation policies (#4181)
The extensions of the BytesScannedWIndowAllocaitonPolicy were stopgap measures to put an allocation policy on the transaction storage. We can now remove that tech debt and reuse the `BytesScannedWindowAllocationPolicy` since the config prefix is decided by the storage key.

**Breaking Changes**

The `metrics` of the allocation policy used to be named the same as the allocation policy class name. That has been changed to always have `allocation_policy` as the name and the rest of the information to be default tags. As a result. The allocation policy dashboard will have to be fixed after this PR is merged