feat: Snuba consumer takes join_timeout as CLI arg (#4192)
If a join timeout is passed, any remaining unprocessed or partially processed messages are dropped. Since they aren't committed yet, we will pick them up again when the consumer restarts.

This is an experiment to see if passing a join timeout fixes the rebalancing issues we are seeing on the generic metrics consumer.