feat(spans): Extract information from span data (#4224)
Upstream now sends relevant data in `spans.data` field of the payload. Use the
information provided in the payload to populate fields in the clickhouse table.
With this, we do not need the logic of performing application logic processing
anymore in snuba spans consumer so we can get rid of it.

In order to validate that the spans processor is performing valid transformations
the PR also adds a test to perform strict validation of the results.