feat(dashboard): Add settings file to frontend code
In order to integrate Sentry into the dashboard frontend, we need a way to have
environment specific settings in the frontend. Add a file that can be overriden
in our production environments.

Also add a Sentry integration that uses that file. This is a noop right now,
but we can have it start working once the appropriate change has been made
in the prod environment.