ref(admin) Allow access to tracing/snql-to-sql by default
Create a new role for product developers that gives access to the tracing and
snql-to-sql tool. Make this role a default role, so everyone with access to the
admin tool will have access to both tools.

Small additional changes: add David to list of users with access to all tools,
and have the product tool role reflect the actual google group we will use to
grant access going forward.