feat: Occasionally log errors if no schema found
We use the validation errors in prod (especially the attached message.txt) to
iterate on the schema. This is kind of a hack to get hold of example
messages before a schema is defined to help speed up the process of creating
a schema for topics where we don't have them yet.