feat(search-issues): Start writing message column in search issues (#4387)
This adds the `message` column to search issues and starts writing to it. `message` is already being
sent, but we missed including it a column initially.

I will let this write for a week or two before switching over to use it. We don't need the full 90
days of data for this to be effective.