feat(snubsplain): Use request context to capture query processors (#4381)
* feat(snubsplain): Use request context to capture query processors

This change is a framework for how meta information about the query execution
will be captured and surfaced through the tool. It starts with one simple data
point, the entity query processors that ran. This can be expanded in subsequent
PRs to add more information.

Since the explain functionality is very similar to what the SnQL-to-SQL tool
already does, the two endpoints were collapsed. The tools are still separate,
and use different parts of the output. In the future the tools can be
 consolidated.

The meta module that was added is the main interface for capturing the query
steps. The data is stored as a list of steps, to preserve the order of execution
during the pipeline.

Changes:
- Add a meta module to provide an interface for capturing steps during query
execution.
- Setup the meta capture based on the `dry_run` flag using the existing
  snql-to-sql tool