feat(capman): add project_id as implicit tenant (#4474)
In order to port the project based rate limiters to the capman framework, we have to have project ids as tenants, we currently do not. Add logic in the build_request function to add project_id to AttributionInfo.tenant_ids

This logic makes an explicit decision that when there are multiple projects being searched in one request, there is no project tenant, instead the organization_id becomes the tenant.