feat: Banish fetching untyped runtime config
User should specify the type when fetching config. Now there are 3 new methods:
get_int_config(), get_float_config() and get_str_config(). get_config() is deprecated
as returning Any is a source of bugs and usually requires the user to cast to the
correct type in their own code anyway.

We also want to call this method from the Rust consumer so we can access runtime config
there, and we should not allow these untyped values there either.