feat(spans): Add migration to fill nodestore version column
We are trying a new scheme for storing span contexts in nodestore. As such
we do not know whether it would cause problems or not. So we need to have
schema version used when storing contexts in nodestore to link things correctly.
Added a migration to add nodestore_column in the spans schema. Added a Delta
encoding since most times we would only have 1 value in the column.