feat(rust-consumer): Handle failed inserts (#4623)
Previously we were treating successful and failed ClickHouse inserts the same, and we simply commit the offset once we get a result from ClickHouse. It's probably better to retry in at least some scenarios. Now the thread will panic. That is probably also not quite right but it's better than silently failing at least for testing purposes.

There was also a bug in the ClickHouse test uncovered by this change. Now we run Snuba migrations in CI before the Rust tests, as it depends on the querylog table being there.