hack(capman): allocation policy violations don't impact SLO (#4642)
The existing SLO calculation depends on the query_metadata_list being updated with the correct request status. Make sure that is done for allocation policy rejections even though no query ever runs. This is not the prettiest way to do it but it unblocks getting rate limiters onto capman and we are redoing how the SLO is calculated anyways