Revert "Revert "hack(capman): allocation policy violations don't impact SLO (#4642)"" (#4662)
This PR was orignially reverted out of caution because when it was deployed the load on a clickhouse replica and the query node went up significantly. With some post-hoc analysis, I cannot find a reason why this would cause the concurrent queries to go up. Here are the reasons why I think these are unrelated:

* This PR only changes the reporting of the queries, not the way queries are executed
* There is no active allocation policy concurrent limiting at the moment and this PR does not change the amount of queries going to clickhouse
* During the load/concurrent query spike there was a huge spike (187 -> 1100) in TCP_TIME_WAIT sockets on the query node

https://app.datadoghq.com/s/FH6-Y3/zak-spc-r24

which was very disproportional to the concurrent queries hitting the node

* On replica 1-5 (which had the load spike and the concurrent query spike), the uncompressed cache hit ratio and the went down to 0 before the big spike

https://app.datadoghq.com/s/FH6-Y3/ep5-jym-j3g

I believe I've done enough due dilligence to try deploying this PR again. If it happens again then I'll do a deeper dive