chore(metric_validation): Entities to query across organizations (#4679)
In order to able to validate the indexed entries in clickhouse metrics and generice metrics clusters with the entries in indexer, we need the ability to run queries across organizations and not have any restrictions on project ids. Created the necessary entities which do not enforce EntityRequiredColumns validation. This PR would be reverted once the validation is complete.