fix(rust-arroyo): Fix off by one error in AnyMessage::committable() (#4695)
There was a bug in the Reduce strategy - since it was passing the seen offsets, rather than the committable offsets (which should be the message offset + 1), it was always committing one behind what it should be. Hopefully this change makes it easier for users of AnyMessage and harder to commit the wrong offset now.