hack(rust-consumer): Add a temporary way to override concurrency CLI values (#4707)
We are going to be testing a lot of different concurrency values with the Rust consumer. Add a way to quickly do this via runtime config to save roundtrips with ops. Obviously this is not actually runtime and needs a redeploy but at least
it is self serve for the team.