perf(rust): Only pause consumer after 1 second of backpressure
Pausing and resuming consumer is a really slow operation since the
rdkafka buffer is discarded. Let's avoid calling this constantly when
we are in a backpressure state by only actually pausing if we've
been waiting in backpressure for more than 1 second.

If this works, we should probably port this to Python Arroyo too.