ci: Run Snuba tests against multiple ClickHouse versions in CI (#4796)
Not all of these are passing right now or blocking for merge/deploy. This change aims to provide more visibility into which features do not work on future ClickHouse versions so we can prioritize fixing these. It also prevents us merging future changes that do not work well in certain versions of ClickHouse that we should support.