feat(capman): Allow concurrent rate limit policies to set max_query_duration_s (#4805)
Currently, state.max_query_duration_s is used by the concurrent rate limiter to know how long a TTL to set in Redis. That value is currently set to 60 seconds while the max request time we allow is 30 seconds. This means that we are potentially storing double the amount of data in redis as we need.

This PR allows us to change that value at runtime so we can experiment with changing the value and saving redis memory.