feat: Mark first migration blocking (#4849)
I don't think there's a valid use case to recreate the table and rerun all migrations without user intervention.

This is intented as a safeguard - the migrations table was inadvertently dropped during a recent incident, and we don't want to run any migrations in that case. This stops any further migration from proceeding.