feat: Make subscription scheduler invalid interval a metric not warning (#4855)
This gets logged all the time but it's considered somewhat normal. It's about to get a lot more common if we move to a "received" timestamp instead of the timestamp from the Kafka topic which is largely monotonic. So let's remove the noisy.

Also moves the Comparable/Interval/InvalidRange types into the subscriptions module since they are fairly useless anywhere else. These haven't changed functionally at all.