feat(cardinality): Cardinalily analysis reporting (#4893)
As we introduce more modules for metrics extraction from spans there is a bigger
need to keep track of how cardinality evolves for different modules. This PR adds
the queries which are used to perform cardinality analysis to a snuba cli command.
The snuba cli command would run the query, dump the results in a CSV file and then
send the file to a slack channel.