ref(subscriptions): Move --delay-seconds from CLI arg to yaml definition
The main motivations for this are:
1. The amount of delay depends on the syncrhonization timestamp used, and
this is defined at the storage level in code. For example if "orig_message_ts"
is used, a longer delay will be applied than if "received_p99" is used, since
received will be set earlier in the pipeline.
2. The same CLI args get applied in all Sentry deployments, and this
makes it easier to keep them in sync
3. Rolling out different values per storage via CLI will probably break
some of our templates and require too much rework.