fix(migrations): Discover's migrations don't reflect our Prod setup.
Our current `discover` migrations don't leave Discover in a working state for our Production setup.  Discover is a `Merge` engine of our `errors` and `transactions` clusters, the current migrations make `discover_dist` a `Distributed` table for `discover_local` which in turn is trying to be a `Merge` table of `errors_local` and `transactions_local`.  Since we don't have `errors_local` and `transactions_local` in the same cluster as `discover_local`, this leaves a new setup in a state where Discover is pointing to permanently empty tables.

This latest (blocking) migration should get new setups to match our existing Production setup, it drops the existing `discover_dist` as a `Distributed` table and replaces it with a `Merge` table pointing to the `errors_dist` and `transactions_dist` tables which are then `Distributed` tables to their respective clusters.