ref(rust): Put consumer in an Arc so it can be shared (#4954)
Initial refactoring so that the consumer can be shared and referenced
within the assignment callbacks. This is necessary so we can do things
like call commit on the consumer during rebalance callbacks.

This PR is based on https://github.com/getsentry/snuba/pull/4936
which had to be reverted as the consumer fully stopped committing
when that was merged. Though this is only the initial refactoring
and does not actually commit during join yet.

At the moment my best guess of why the previous attempt did not work
in prod was the the context was being mutated after the consumer
was initially created. This version avoids doing this and keeps the
original approach of deferring consumer creation until
StreamProcessor.subscribe() is called.