feat(rust): Port the buffered messages class to Rust
This is the start of the DLQ implementation. It adds the struct
that keeps track of seen messages from the broker so they can be
produced to the DLQ if needed later.
It is basically a 1:1 port of current Arroyo code here (
https://github.com/getsentry/arroyo/blob/112667abdc2ef5c9137e361c0b2bbf69703bc524/arroyo/dlq.py#L215-L269)
that we run in the Python consumer. Actual DLQing of messages
will be implemented as a follow up.