fix(rust): Rename metrics namespace, and implement one metric from python
In order to make the Rust consumer emit the same metrics as Python
consumer, we need to use the same namespace. This is probably risky as
it could break alerts that do not apply the correct tag filters. We'll
just find out, I'm oncall anyway.

Also add a metric that already exists in Python, to see if it will show
up correctly.