feat(rust): Start buffering messages for DLQ
This is a very naive implementation of buffering messages that are
pending commit. Later, it can be used by the DLQ to retrieve
untransformed messages which are deemed invalid and produced to the
DLQ.

I'm adding this functionality now (even though rest of DLQ is not really
done) because:
1. I think it might have some implications for how we've defined TPayload
in a bunch of places (it probably needs to be Clone), and
2. It is likely to have performance implications that we should start
understanding better now.

Even if we try to make message cheaper to clone (there was some discussion
about using an Arc internally for the payload), I'm not sure there's any way
around having the payload be Clone.

Might need to bump requested memory a bit when we roll this out.