fix(rust): Fix the ProduceCommitLog strategy
The ProduceCommitLog gets a message with type BytesInsertBatch
from the prior processing step, and produces one message for each
partition contained within that batch.

In order to achieve this, this splits the `BytesInsertBatch` into
two separate structs. Now we have `RowData` which we require the
message processors to return. `BytesInsertBatch` is internal and
contains other metadata that we will add for producing to the commit
log.