fix(rust-python): Join handles less frequently, fix backpressure bug
max_queue_depth does not work as it is implemented right now, but now
that we expose it as a parameter, we might as well use it to determine
how many handles are supposed to be in-flight.

The original theory was correct though: We are blocking too much in
check_for_results, and that's the culprit for our performance problems.

Now the next performance problem is the serialization overhead for
individual messages between processes, and we might have to implement
batching to fix that. But at the very least the processor can now
saturate 5 cores on my machine.