Forward rdkafka logging
The Rust bindings should be doing this out of the box, but we will still implement this manually, to emit logs from `rust_arroyo` rather, and to properly use a proper `&dyn std::error::Error` which Sentry can capture.

As a driveby change, this also removed a ton of unused dependencies.