feat(rust): Add --no-strict-offset-reset to the Rust CLI (#5163)
This doesn't do anything yet, just adding the flag first and passing it through to the Rust Kafka config. 
Merging this first allows us to start passing this flag in some of our test consumers that are already running before we flip the actual behavior over to strict mode with the rest of the implementation.