feat(mql): Add MQL to Snuba AST parser (#5119)
Add a parser that converts from an MQL string and a context dictionary into a Snuba AST query.

This parser only deals with `LogicalQuery` objects, since we don't currently support `CompositeQuery` in MQL queries.

Blast Radius:
Zero. This code isn't exposed in any endpoint at the moment. Further PRs will wire this up to an API endpoint.

Notes:

The `MQLContext` class is very close to the `snuba_sdk.MQLContext` object. However there was a decision to not reuse that class in Snuba. The main reason was to keep Snuba uncoupled from the Python SDK. One way to think about this is that (in theory) there could be more than one SDK (Go? Rust?) that Snuba has to handle. Coupling to one specific SDK is an incorrect way to structure our code.

This PR has one large implication for Sentry: MQL doesn't allow for sending raw metric IDs. That means Sentry has to send the MRI string. Since currently only Sentry has access to the indexer, that means Sentry has to look up the resolutions of the strings, store those look-ups, and send them through to Snuba so Snuba can replace them with the appropriate IDs. At a future date Snuba should be able to do the resolution directly.

---------

Co-authored-by: Evan Hicks <evan.hicks@sentry.io>