feat(rust): Align offset reset handling to the Python implementation (#5164)
This aligns the offset reset handling with the Python version of Arroyo, which is stricter than how things work in the underlying rdkafka library. Offsets are explicitly assigned during the assignment callback and will not be automatically reset unless there is no initial offset in the store (new consumer group).

The behavior can be overridden (essentially restoring the original rdkafka behavior) by setting `strict_offset_reset` to False (true is the default). This is only advised for dev like environments, as we do not want to be inadvertently skipping messages in production.