fix: Auto offset reset when no initial offset is stored
Currently there is a bug where if no initial offset is stored
the consumer will just crash with `KafkaError (Message consumption error: AutoOffsetReset (Local: No offset to automatically reset to`

This is just a demo of how calling consumer::assign() outside
of the rebalance callback actually works on initial assignment
fixes this. This code is an abomination and shouldn't actually
be shipped to production.