fix(rust): Do not drop runtime and accidentally cancel tasks
We were not producing anything to the DLQ because after calling the
async fn, we scheduled the future onto a runtime that we already dropped
in the constructor of DlqProducer. Holding onto the Handle is not
enough, we need the entire Runtime.

Also adjust tests so the error becomes apparent: now the counter is not
incremented if the future isn't polled at least once.