fix(spans): Add sentry_tags to spans entity (#5162)
The spans data model has 2 sets of tags
1. Customer sent tags.
2. Sentry extracted tags.

We are duplicating the sentry extracted tags into customer sent tags because product wasn’t aware of this distinction. But we are having a storage cost impact of that. By exposing the sentry extracted tags in the entity we can stop duplicating data and get rid of the redundancy.