perf: Increase consumer join timeout (#5200)
A few consumers (spans, generic metrics distributions) regularly exceed the 5 second timeout currently configured, and cannot finish processing pending work in time. When this is the case, the multiprocessing pool cannot be reused since there is state in the pool that cannot be carried over to the next assignment. This aims to avoid that scenario.